requirements.txt
# Core Dependencies
python-telegram-bot[job-queue]==20.7
pymongo==4.6.1
python-dotenv==1.0.0
aiohttp==3.9.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Optional but Recommended
# Pillow==10.1.0  # For advanced image processing
# pytesseract==0.3.10  # For OCR text extraction from images